# pkg/openapi/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package api

import (
	"net/http"
	"net/url"
	"strings"

	"github.com/ogen-go/ogen/uri"
)

var (
	rn1AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
	rn3AllowedHeaders = map[string]string{
		"DELETE": "Content-Type",
	}
	rn4AllowedHeaders = map[string]string{
		"PATCH": "Content-Type",
	}
	rn6AllowedHeaders = map[string]string{
		"GET": "Content-Type",
	}
	rn5AllowedHeaders = map[string]string{
		"GET": "Content-Type",
	}
)

func (s *Server) cutPrefix(path string) (string, bool) {
	prefix := s.cfg.Prefix
	if prefix == "" {
		return path, true
	}
	if !strings.HasPrefix(path, prefix) {
		// Prefix doesn't match.
		return "", false
	}
	// Cut prefix from the path.
	return strings.TrimPrefix(path, prefix), true
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	elemIsEscaped := false
	if rawPath := r.URL.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
			elemIsEscaped = strings.ContainsRune(elem, '%')
		}
	}

	elem, ok := s.cutPrefix(elem)
	if !ok || len(elem) == 0 {
		s.notFound(w, r)
		return
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/task/"

			if l := len("/task/"); len(elem) >= l && elem[0:l] == "/task/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'c': // Prefix: "create"

				if l := len("create"); len(elem) >= l && elem[0:l] == "create" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "POST":
						s.handleHandleCreateTaskRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "POST",
							allowedHeaders: rn1AllowedHeaders,
							acceptPost:     "application/json",
							acceptPatch:    "",
						})
					}

					return
				}

			case 'd': // Prefix: "delete"

				if l := len("delete"); len(elem) >= l && elem[0:l] == "delete" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "DELETE":
						s.handleHandleDeleteTaskRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "DELETE",
							allowedHeaders: rn3AllowedHeaders,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}

			case 'f': // Prefix: "finish"

				if l := len("finish"); len(elem) >= l && elem[0:l] == "finish" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "PATCH":
						s.handleHandleFinishTaskRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "PATCH",
							allowedHeaders: rn4AllowedHeaders,
							acceptPost:     "",
							acceptPatch:    "application/json",
						})
					}

					return
				}

			case 'g': // Prefix: "get"

				if l := len("get"); len(elem) >= l && elem[0:l] == "get" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch r.Method {
					case "GET":
						s.handleHandleGetTaskRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET",
							allowedHeaders: rn6AllowedHeaders,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}
				switch elem[0] {
				case '_': // Prefix: "_list"

					if l := len("_list"); len(elem) >= l && elem[0:l] == "_list" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "GET":
							s.handleHandleGetListOfTasksRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "GET",
								allowedHeaders: rn5AllowedHeaders,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}

				}

			}

		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name           string
	summary        string
	operationID    string
	operationGroup string
	pathPattern    string
	count          int
	args           [0]string
}

// Name returns ogen operation name.
//
// It is guaranteed to be unique and not empty.
func (r Route) Name() string {
	return r.name
}

// Summary returns OpenAPI summary.
func (r Route) Summary() string {
	return r.summary
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.operationID
}

// OperationGroup returns the x-ogen-operation-group value.
func (r Route) OperationGroup() string {
	return r.operationGroup
}

// PathPattern returns OpenAPI path.
func (r Route) PathPattern() string {
	return r.pathPattern
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
//
// Note: this method does not unescape path or handle reserved characters in path properly. Use FindPath instead.
func (s *Server) FindRoute(method, path string) (Route, bool) {
	return s.FindPath(method, &url.URL{Path: path})
}

// FindPath finds Route for given method and URL.
func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
	var (
		elem = u.Path
		args = r.args
	)
	if rawPath := u.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
		}
		defer func() {
			for i, arg := range r.args[:r.count] {
				if unescaped, err := url.PathUnescape(arg); err == nil {
					r.args[i] = unescaped
				}
			}
		}()
	}

	elem, ok := s.cutPrefix(elem)
	if !ok {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/task/"

			if l := len("/task/"); len(elem) >= l && elem[0:l] == "/task/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'c': // Prefix: "create"

				if l := len("create"); len(elem) >= l && elem[0:l] == "create" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "POST":
						r.name = HandleCreateTaskOperation
						r.summary = "Добавляет новую задачу"
						r.operationID = "HandleCreateTask"
						r.operationGroup = ""
						r.pathPattern = "/task/create"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			case 'd': // Prefix: "delete"

				if l := len("delete"); len(elem) >= l && elem[0:l] == "delete" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "DELETE":
						r.name = HandleDeleteTaskOperation
						r.summary = "Удалить задачу"
						r.operationID = "HandleDeleteTask"
						r.operationGroup = ""
						r.pathPattern = "/task/delete"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			case 'f': // Prefix: "finish"

				if l := len("finish"); len(elem) >= l && elem[0:l] == "finish" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "PATCH":
						r.name = HandleFinishTaskOperation
						r.summary = "Завершить задачу"
						r.operationID = "HandleFinishTask"
						r.operationGroup = ""
						r.pathPattern = "/task/finish"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			case 'g': // Prefix: "get"

				if l := len("get"); len(elem) >= l && elem[0:l] == "get" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "GET":
						r.name = HandleGetTaskOperation
						r.summary = "Получить задачу"
						r.operationID = "HandleGetTask"
						r.operationGroup = ""
						r.pathPattern = "/task/get"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case '_': // Prefix: "_list"

					if l := len("_list"); len(elem) >= l && elem[0:l] == "_list" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "GET":
							r.name = HandleGetListOfTasksOperation
							r.summary = "Получить список задач"
							r.operationID = "HandleGetListOfTasks"
							r.operationGroup = ""
							r.pathPattern = "/task/get_list"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				}

			}

		}
	}
	return r, false
}